+++ app/main/views/feedback.py
@@ -1,0 +1,42 @@
+import markdown
+import requests
+from flask import render_template, url_for, redirect, flash, current_app, abort
+from app.main import main
+from flask_login import login_required
+from app.main.forms import Feedback
+
+from flask.ext.login import current_user
+
+
+@main.route('/feedback', methods=['GET', 'POST'])
+def feedback():
+    form = Feedback()
+    if form.validate_on_submit():
+        data = {
+            'person_email': current_app.config.get('DESKPRO_PERSON_EMAIL'),
+            'department_id': current_app.config.get('DESKPRO_TEAM_ID'),
+            'subject': 'Notify feedback',
+            'message': '{}\n{}\n{}'.format(
+                form.name.data,
+                form.email_address.data,
+                form.feedback.data)
+        }
+        headers = {
+            "X-DeskPRO-API-Key": current_app.config.get('DESKPRO_API_KEY'),
+            'Content-Type': "application/x-www-form-urlencoded"
+        }
+        resp = requests.post(
+            current_app.config.get('DESKPRO_API_HOST') + '/api/tickets',
+            data=data,
+            headers=headers)
+        if resp.status_code != 201:
+            current_app.logger.error(
+                "Deskpro create ticket request failed with {} '{}'".format(
+                    resp.status_code,
+                    resp.json())
+                )
+            abort(500, "Feedback submission failed")
+        flash("Your feedback has been submitted")
+        return redirect(url_for('.feedback'))
+
+    return render_template('views/feedback.html', form=form)
